+++ packages/web-components/src/components/wysiwyg/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/wc-wysiwyg@0.1.0...@carbon-labs/wc-wysiwyg@0.2.0) (2026-09-21)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @carbon/element-styles to ^0.3.0 ([#1376](https://github.com/carbon-design-system/carbon-labs/issues/1376)) ([c2f186c](https://github.com/carbon-design-system/carbon-labs/commit/c2f186c098c78c327052e9c832c33cc5f723d318))
+
+
+
+
+
 # 0.1.0 (2026-07-20)